Remove BlobMeta
