+++ src/hrb/BlobDatabase.hh
@@ -48,6 +48,7 @@ public:
 	BlobResponse response(
 		ObjectID id,
 		unsigned version,
+		std::string_view mine,
 		std::string_view etag,
 		std::string_view rendition = {}
 	) const;
