+++ src/hrb/Ownership.cc
@@ -14,13 +14,7 @@
 #include "Ownership.ipp"
 #include "BlobDatabase.hh"
 
-#include <rapidjson/document.h>
-#include <rapidjson/ostreamwrapper.h>
-#include <rapidjson/pointer.h>
-#include <rapidjson/writer.h>
-
 #include <sstream>
-#include <cassert>
 
 namespace hrb {
 
@@ -129,67 +123,6 @@ std::string Ownership::Collection::serialize(redis::Reply& reply) const
 	});
 	ss << "}}";
 	return ss.str();
-}
-
-CollEntry::CollEntry(std::string_view redis_reply) : m_raw{redis_reply}
-{
-}
-
-std::string CollEntry::create(std::string_view perm, std::string_view filename, std::string_view mime)
-{
-	rapidjson::Document json;
-	json.SetObject();
-	json.AddMember("mime", rapidjson::StringRef(mime.data(), mime.size()), json.GetAllocator());
-
-	if (!filename.empty())
-		json.AddMember("filename", rapidjson::StringRef(filename.data(), filename.size()), json.GetAllocator());
-
-	assert(perm.size() == 1);
-	std::ostringstream ss;
-	ss << perm;
-
-	rapidjson::OStreamWrapper osw{ss};
-	rapidjson::Writer<rapidjson::OStreamWrapper> writer{osw};
-	json.Accept(writer);
-	return ss.str();
-}
-
-std::string_view CollEntry::json() const
-{
-	auto json = m_raw;
-	json.remove_prefix(1);
-	return json;
-}
-
-std::string CollEntry::filename() const
-{
-	auto json = CollEntry::json();
-
-	rapidjson::Document doc;
-	doc.Parse(json.data(), json.size());
-	if (doc.HasParseError())
-		return {};
-
-	auto&& val = GetValueByPointerWithDefault(doc, "/filename", "");
-	return std::string{val.GetString(), val.GetStringLength()};
-}
-
-std::string CollEntry::mime() const
-{
-	auto json = CollEntry::json();
-
-	rapidjson::Document doc;
-	doc.Parse(json.data(), json.size());
-	if (doc.HasParseError())
-		return {};
-
-	auto&& val = GetValueByPointerWithDefault(doc, "/mime", "");
-	return std::string{val.GetString(), val.GetStringLength()};
-}
-
-bool CollEntry::allow(std::string_view user) const
-{
-	return Permission{m_raw.substr(0, 1)}.allow(user);
 }
 
 } // end of namespace hrb
